Move Array helper code to ArrayUtil
