+++ python/python/Ice/__init__.py
@@ -196,53 +196,3 @@ UnknownSlicedValue._ice_type = IcePy._t_UnknownSlicedValue
 Protocol_1_0 = Ice.ProtocolVersion(1, 0)
 Encoding_1_0 = Ice.EncodingVersion(1, 0)
 Encoding_1_1 = Ice.EncodingVersion(1, 1)
-
-
-BuiltinBool = 0
-BuiltinByte = 1
-BuiltinShort = 2
-BuiltinInt = 3
-BuiltinLong = 4
-BuiltinFloat = 5
-BuiltinDouble = 6
-
-BuiltinTypes = [
-    BuiltinBool,
-    BuiltinByte,
-    BuiltinShort,
-    BuiltinInt,
-    BuiltinLong,
-    BuiltinFloat,
-    BuiltinDouble,
-]
-BuiltinArrayTypes = ["b", "b", "h", "i", "q", "f", "d"]
-
-
-def createArray(view, t, copy):
-    if t not in BuiltinTypes:
-        raise ValueError("`{0}' is not an array builtin type".format(t))
-    a = array.array(BuiltinArrayTypes[t])
-    a.frombytes(view)
-    return a
-
-
-try:
-    import numpy
-
-    BuiltinNumpyTypes = [
-        numpy.bool_,
-        numpy.int8,
-        numpy.int16,
-        numpy.int32,
-        numpy.int64,
-        numpy.float32,
-        numpy.float64,
-    ]
-
-    def createNumPyArray(view, t, copy):
-        if t not in BuiltinTypes:
-            raise ValueError("`{0}' is not an array builtin type".format(t))
-        return numpy.frombuffer(view.tobytes() if copy else view, BuiltinNumpyTypes[t])
-
-except ImportError:
-    pass
